Add mailmap file to normalize author strings
This commit adds a mailmap file to normalize the author strings in the
git log. Some of the email addresses used in the git log were different
than the preferred addresses used by the commit authors. The mailmap
file gives a place to normalize the author strings in the git log.